requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langchain_chroma
chromadb
langgraph
requests
beautifulsoup4
aiohttp
python-dotenv